Added an option of words in a query that trigger args to be redacted
The idea here is to not log sensitive info. The assumption is that they
will be passed as parameters to a prepared statement, but the idea
could be expanded to redact the query, too, in case they are already
interpolated in.

This adds an O(n) loop over all the redaction triggers along with a
substring check on each log statement.

I had wanted to use a set of triggers and check each NamedValue's Name,
but apparently the sqlserver driver (at least), leaves Name empty.
:shrug: